requirements.txt
aiofiles==25.1.0
aiogram==3.26.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
asyncpg==0.31.0
attrs==25.4.0
certifi==2026.2.25
colorama==0.4.6
colorlog==6.10.1
dotenv==0.9.9
frozenlist==1.8.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
magic-filter==1.0.12
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dotenv==1.2.2
sqlparse==0.5.5
tabulate==0.10.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.23.0
yoyo-migrations==9.0.0
zipp==3.23.0
